Fix data type in ResponsePage

+++ passenger-service/src/main/java/com/kpavlov/modulepassengerservice/dto/response/PassengerResponsePage.java
@@ -13,5 +13,5 @@ public record PassengerResponsePage (
 
         int totalPages,
 
-        long limit
+        int limit
 ) {}
